backend/requirements.txt
# Production RAG Backend Dependencies
# Core FastAPI and Async Services
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
deprecated==1.2.18
dirtyjson==1.0.8
dnspython==2.8.0
ecdsa==0.19.1
eval-type-backport==0.2.2
eventlet==0.40.3
fastapi==0.115.6
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.4.0
google-api-core==2.26.0
google-auth==2.41.1
google-generativeai==0.3.2
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.62.3
grpcio-tools==1.62.3
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
joblib==1.5.2
# Updated LlamaIndex packages for production stability
llama-index-core==0.12.23
llama-index-embeddings-together==0.3.0
llama-index-readers-file==0.4.0
llama-index-vector-stores-qdrant==0.4.0
llama-index-llms-together==0.3.0
llama-index-readers-web==0.3.0
markdown-it-py==4.0.0
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.7.0
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==1.26.4
orjson==3.9.10
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pdfminer-six==20221105
pdfplumber==0.10.3
pillow==10.4.0
pluggy==1.6.0
portalocker==2.10.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.11.0
pygments==2.19.2
pypdf==4.3.1
pypdf2==3.0.1
pypdfium2==4.30.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
pywin32==311
pyyaml==6.0.3
qdrant-client==1.13.3
regex==2025.10.23
requests==2.32.5
rich==13.7.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.44
starlette==0.41.3
striprtf==0.0.26
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.12.0
together==1.2.11
tqdm==4.67.1
typer==0.12.5
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.32.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0

# Production Async Services and Task Management
redis==5.2.1
redis-py-cluster==2.1.3
celery==5.4.0
kombu==5.4.2
billiard==4.2.1
vine==5.1.0
amqp==5.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0

# Circuit Breaker and Resilience
circuitbreaker==2.0.0

# LangChain Production Packages
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.5
langchain-openai==0.2.14
langchain-redis==0.2.1

# Additional Production Monitoring and Logging
structlog==24.4.0
python-json-logger==3.2.1

# Security and Performance
python-telegram-bot==21.7
cryptography==46.0.3
